move gui specific code to gui_util from search class
