+++ src/search_ui.h
@@ -32,7 +32,6 @@ private:
 	wxBitmap GetBitmapForExtension(std::string ext);
 	void OnCheck(wxCommandEvent& event);
 	void OnCheck(wxCheckBox* checkbox);
-	void HighlightWord(std::string);
 	void SetColor(wxColour, int);
 
 	void NewSearch(std::string, std::string);
