+++ src/util/word_util.h
@@ -1,0 +1,9 @@
+#ifndef WORD_UTIL_H_
+#define WORD_UTIL_H_
+
+namespace word_util {
+
+
+}
+
+#endif // WORD_UTIL_H_
